+++ src/code/util/debug/_.rs
@@ -3,15 +3,18 @@
 #![doc = crate::_DOC_CODE_UTIL_DEBUG!()] // public
 #![doc = crate::_doc!(modules: crate::code::util; debug)]
 #![doc = crate::_doc!(flat:"code")]
-#![doc = crate::_doc!(hr)]
+#![doc = crate::_doc!(extends: backtrace)]
 //
 
 crate::mods_in! {
+    #[cfg(feature = "std")]
+    mod _reexport_std;
+
     mod cdbg;
     mod fn_name;
     mod warn;
 }
-crate::mods_out! { // _mods
+crate::mods_out! { // _mods, _reexports
     _mods {
         #[doc(inline)]
         pub use super::{
@@ -19,5 +22,9 @@ crate::mods_out! { // _mods
             fn_name::fn_name,
             warn::const_warn,
         };
+    }
+    _reexports {
+        #[cfg(feature = "std")]
+        pub use super::_reexport_std::*;
     }
 }