+++ src/error/data/_.rs
@@ -1,6 +1,9 @@
 // devela/src/error/data/_.rs
 //
-//! Data-related errors.
+#![doc = crate::_DOC_ERROR_DATA!()] // public
+#![doc = crate::_doc!(modules: crate::error; data)]
+#![doc = crate::_doc!(flat:"error")]
+#![doc = crate::_doc!(hr)]
 //
 
 crate::mods_in! {