+++ src/error/num/_.rs
@@ -2,7 +2,7 @@
 //
 #![doc = crate::_DOC_ERROR_NUM!()] // public
 #![doc = crate::_doc!(modules: crate::error; num)]
-#![doc = crate::_doc!(flat:"num")]
+#![doc = crate::_doc!(flat:"error")]
 #![doc = crate::_doc!(hr)]
 //