+++ docs/changelog.md
@@ -25,5 +25,13 @@
 
 # Modules
 
+### code::util::assert
+- move `compile_error!` from `error`.
+
+### code::util::debug
+- move `Backtrace` and `BacktraceStatus` from `error`.
+
+## error
+- make `kind` public.
 
 [0.30.0]: https://github.com/andamira/devela/releases/tag/v0.30.0